Moved feature density functions to seperate file, added start of horizon line functions
