+++ Source/boundingBoxes.py
@@ -10,8 +10,8 @@ import cv2
 from numba import jit, njit
 
 # Custom  imports
-from Source.logger import Logger
-from Source import exceptions
+from logger import Logger
+import exceptions
 
 
 # gets the coordinates out of the bounding box list/array
@@ -28,7 +28,7 @@ def getBoundingBoxPoints(box):
 #        [ [x1, y1], [x2, y2] ]
 #        where x1, y1, x2, y2 are any number
 # the number type gets sanitized upon boundingBox load coordinates
-def drawBoundingBoxes(rawImage, boundingBoxes, color=(0,0,255), thickness=2, windowName="Bounding Boxes"):
+def drawBoundingBoxes(rawImage, boundingBoxes, color=(0,0,255), thickness=2, windowName="Bounding Boxes", show=False):
     image = np.copy(rawImage)
     for box in boundingBoxes:
         p1, p2, p3, p4 = getBoundingBoxPoints(box)
@@ -36,4 +36,6 @@ def drawBoundingBoxes(rawImage, boundingBoxes, color=(0,0,255), thickness=2, win
         cv2.line(image, p2, p3, color, thickness)
         cv2.line(image, p3, p4, color, thickness)
         cv2.line(image, p4, p1, color, thickness)
+    if show:
         cv2.imshow(windowName, image)
+    return image
